+++ CHANGELOG.md
@@ -12,7 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Add admin component SignIn
 - Add admin component SignUp
 - Add admin component AdminHeader with logout
-- Add demo admin component SideMenu
+- Add admin component SideMenu, with new routes using uiMenu
 - Add demo admin component DashboardAdmin
 - Add basic store pages
 - Add basic componnet ListPages